requirements.txt
django==5.1.3

appleauth==0.0.1
django-environ==0.11.2
djangorestframework==3.15.2
PyJWT==2.10.0
requests==2.32.3
